move map plot utils to separate file
